refactor: Extract inline tests to external modular files
- Removed all #[cfg(test)] inline tests from main.rs and config.rs
- Created modular test structure following CLAUDE.md requirements:
  - src/tests/main/ - CLI argument and command parsing tests (28 tests)
  - src/tests/config/ - Configuration loading and management tests (11 tests)
- All tests have descriptive two-line comments explaining purpose
- Test utilities properly separated into dedicated test_utils.rs files
- Maintains 90%+ test coverage requirement (90.84% line coverage)
- Follows established pattern from core/executor and core/plugin tests
