+++ src/tests.rs
@@ -1,0 +1,2 @@
+mod config;
+mod main;
