+++ src/tests/config/mod.rs
@@ -1,0 +1,5 @@
+pub mod defaults;
+pub mod hooks;
+pub mod loading;
+pub mod serialization;
+pub mod test_utils;
