+++ src/tests/core_hooks/mod.rs
@@ -1,0 +1,5 @@
+pub mod actions;
+pub mod context;
+pub mod execution;
+pub mod patterns;
+pub mod test_utils;
