+++ src/tests/main/mod.rs
@@ -1,0 +1,5 @@
+pub mod argument_parsing;
+pub mod command_parsing;
+pub mod configuration;
+pub mod test_utils;
+pub mod utility_flags;
